Try to start merging leader concept into escort system
Remove p->leader, and set p->parent and p->escorts
