+++ src/escort.h
@@ -14,6 +14,7 @@
 
 /* Creation. */
 int escort_addList( Pilot *p, char *ship, EscortType_t type, unsigned int id );
+int escort_rmList( Pilot *p, unsigned int id );
 unsigned int escort_create( Pilot *p, char *ship,
       Vector2d *pos, Vector2d *vel, double dir,
       EscortType_t type, int add );
